Fix shared permission policy boundaries
Copilot-Session: ec9b12fc-e82b-4baf-b692-abb37c7d2e8f
